code/requirements.txt
absl-py==1.0.0
adamp==0.3.0
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.4
cryptography @ file:///tmp/build/80754af9/cryptography_1639414572950/work
datasets==1.18.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
docker-pycreds==0.4.0
easydict==1.9
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
filelock==3.6.0
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
huggingface-hub==0.4.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.11.3
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
multidict==6.0.2
multiprocess==0.70.12.2
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
nltk==3.7
notebook==6.4.10
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095651905/work
oauthlib==3.2.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyDeprecate==0.3.1
Pygments==2.11.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
regex==2022.3.15
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.8
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
sentry-sdk==1.5.8
setproctitle==1.2.2
shortuuid==1.0.8
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.1
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.7.2
torchvision==0.12.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
transformers==4.17.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
wandb==0.12.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.0
xxhash==3.0.0
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
